SAA-1055: Improved unit tests for async
